add function that raises an error if output file(s) already exist, and add a test for this
