+++ tests/test_tools.py
@@ -847,3 +847,32 @@ def test_load_metadata(mocker, user_filesystem, inputs, expected):
             **expected,
         }
         assert actual_metadata == expected_metadata
+
+
+def test_preprocess_args_bad(user_filesystem):
+    # Case: user tries to run absorption correction, but the output
+    # filenames already for *_corrected.chi and *_cve.chi exists.
+    # expected: preprocess_args catches this early and raises an Error
+    input_data_file = str(user_filesystem / "good_data.chi")
+    existing_corrected_file = str(
+        user_filesystem / "output_dir" / "good_data_corrected.chi"
+    )
+    existing_corrected_cve_file = str(
+        user_filesystem / "output_dir" / "good_data_cve.chi"
+    )
+    cli_inputs = (
+        ["mud"]
+        + [input_data_file]
+        + ["2.5", "-w", "Mo", "-o", str(user_filesystem / "output_dir")]
+        + ["-c"]  # -c flag saves the cve file
+    )
+    args = get_args_cli(cli_inputs)
+    args = set_input_lists(args)
+    msg = (
+        "The following output files already exist:"
+        f"\n{existing_corrected_file}\n"
+        f"{existing_corrected_cve_file}\n"
+        "Use --force to overwrite them."
+    )
+    with pytest.raises(FileExistsError, match=re.escape(msg)):
+        preprocessing_args(args)
